Modified header file by adding struct and alias (typedef)

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,11 +1,14 @@
+#include "lists.h"
+
 /**
- * print_list - This is a function that prints all the elements of a list_t list
+ * print_list - This is a function that prints all the
+ * elements of a list_t list
  * @h: Pointer to the list
  * Return: Returns number of nodes
  */
 size_t print_list(const list_t *h)
 {
-	size_t counter = 0;
+	unsigned int counter = 0;
 
 	while (h != NULL)
 	{
